requirement.txt
Flask
google-cloud-firestore==1.6.2
firebase-admin